+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `table` output mode
 - Filter repositories by group
 
+### Changed
+
+- `--name` parameter for repositories filtering to consume regexp
+
 ## [0.4.0] - 2022-07-29
 
 ### Added